requirements.txt
aiogram==2.25.1
paho-mqtt==1.6.1
requests==2.31.0
